requirements.txt
django==4.1.2
pytest==7.1.3
pytest-cov==4.0.0
pylint==2.15.4
black==22.10.0
fire==0.4.0
dj-database-url==1.0.0
python-decouple==3.6
collectfast==2.2.0
django-mathfilters==1.0.0
django-rest_framework==0.1.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
psycopg2-binary==2.9.4
django-admin-rangefilter==0.9.0
pytest-django==4.5.2
boto3==1.24.89
sentry-sdk==1.9.10
django-s3-folder-storage==0.5
PyJWT==2.5.0